+++ src/Set/Provider/Strings/Unicode.php
@@ -3191,7 +3191,7 @@ final class Unicode implements Provider
      *
      * @template V
      *
-     * @param callable(string): (V|Seed<V>) $map
+     * @param callable(string): V $map
      *
      * @return Set<V>
      */